Remove redundant EngineTest project
Deleted the redundant EngineTest project and the corresponding VC project file as it is not needed for the main project operation. This was done to declutter the codebase and keep only relevant code. The EngineTest project was initially setup for testing purposes which are no longer needed. Also, irrelevant lines have been removed from the solution (oloprox.sln) and insignificant changes have been made in main.cpp to adjust to this removal.