Update oloprox.vcxproj, main.cpp, and config files
In this commit, the necessary modifications have been applied to the "oloprox.vcxproj" and "main.cpp" files to add the <BuildStlModules> setting. It ensures that the modules of the Standard Template Library are built. The source path in "config.yml" and "CominingBuildingFile/config.yml" files are also updated to fix the broken references.

A new logic has been added to debug the file action and calculate SHA-256 hash of a file in "main.cpp". This helps in tracking file changes, especially, in a collaborate work environment.

In README.md, the static content about a 2D platformer project description is replaced with a simple project title and a thank you note. Furthermore, boost filesystem library and spdlog logging library have been integrated in the "main.cpp" file to provide advanced filesystem operations and logging functions.

The "config.yml" file in "build tools" folder has been modified to include the binary type of the build, allowing the user to specify whether to perform a debug or release build.