Refactor and update observer pattern, file-copy and ignore list functionalities
This commit focuses on modifying the observer pattern and file copy functionalities. The ignore list now not only includes file extensions, but also directories and files. The details of the file hash check and copy are moved into deteils.hpp for a more organized structure of the code. Some unnecessary codes are trimmed to make the code cleaner. A new file, deteils.hpp, is added; changes are also made to .idea.oloprox/sonarlint/issuestore/index.pb, config.yml and main.cpp.

The purpose of this commit is to optimize code structure for better maintenance, and to enhance the efficiency and convenience in dealing with large numbers of files through modifications in ignore list and file-copy functionalities.