Refactor file handling and simplify catch block
This commit includes several changes that simplifies file handling and exceptions. Unused includes were removed in the oloprox module while the file handling methods in the utils namespace were simplified and moved to a new filesystem namespace. The catch block in system.ixx was also simplified to catch any kind of exception.