Remove old configuration and synchronization classes, add new structure
Multiple old classes for file synchronization and configuration were deleted from the code. In replacement, a new structure in including the 'create_logger' function and the 'Configer' class was implemented in the 'utils' namespace. File paths were updated and some minor fixes were made in different parts of the code for optimization and error prevention purposes.